% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electricityDemand.R
\docType{data}
\name{electricityDemand}
\alias{electricityDemand}
\title{Peninsular Spanish Electricity Demand}
\format{
An object of class \code{"fds"}.
}
\usage{
data("electricityDemand")
}
\description{
Electricity demand each 10 minutes from 2014 to 2018.
}
\references{
\url{http://www.ree.es/}
}
\keyword{datasets}
