% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{envelope}
\alias{envelope}
\title{Envelope Algorithm}
\usage{
envelope(data, focal, distance, plot)
}
\arguments{
\item{data}{a fts or fds object see R package rainbow or ftsa.}

\item{focal}{the name of the curve to envelope. In the article, the most recent curve.}

\item{distance}{a character string naming the distance. Available options are "supremum" and "l2".}

\item{plot}{if plotting the selected curves in the different iterations of the algorithm.}
}
\value{
a list containing the selected curves.
}
\description{
This function implement the algorithm to optain neighbourhoods as in the paper.
}
\examples{
data(sinedata)
focal <- '1'
resultsBand <- envelope(sinedata, focal, distance = "l2", plot = FALSE)

}
