% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday.R
\docType{data}
\name{holiday}
\alias{holiday}
\title{Holidays in Spain}
\format{
An object of class \code{"list"}.
}
\usage{
data("holiday")
}
\description{
A list containing the bank holidays in Spain from 2014 to 2018.
}
\keyword{datasets}
