% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBand.R
\name{plotBand}
\alias{plotBand}
\title{Plot Bands}
\usage{
plotBand(data, focal, Jordered, kcurves)
}
\arguments{
\item{data}{a fts or fds object see R package rainbow or ftsa.}

\item{focal}{the name of the curve to envelope. In the article, the most recent curve.}

\item{kcurves}{number of curves to build the band.}

\item{Joredered}{set of curves ordered by distance or depth.}
}
\value{
a list containing the selected curves.
}
\description{
This function plot the band given by the k deepest or closest functions.
}
