% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinedata.R
\docType{data}
\name{sinedata}
\alias{sinedata}
\title{Simulated data}
\format{
An object of class \code{"fds"}.
}
\usage{
data("sinedata")
}
\description{
Functional data generated as the sum of a cosine and sine functions with random coefficients.
}
\keyword{datasets}
