% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{envelope}
\alias{envelope}
\title{envelope}
\usage{
envelope(data, focal, dist = "l2", plot = TRUE)
}
\arguments{
\item{data}{Do you love cats? Defaults to TRUE.}

\item{focal}{Column name of the function to envelope.}

\item{dist}{Distance between functions ("l2" or "supremum"). Defaults is L2 distance.}

\item{plot}{If plotting or not the iterations. Defaults to TRUE.}
}
\description{
This function allows to get the envelopes.
}
\examples{
envelope()
}
\keyword{envelope}
